Remove RemainAfterExit in systemd unit
RemainAfterExit means that the service should be considered active even
when all its children has terminated. This is not what we want here as
the child process created by carton is a daemon process and not expected
to terminate.

Also, on stretch, the return code of the child process has no influence
on the status of the service if RemainAfterExit is set.